Conflicts:
	plugins/plugins.pro
	plugins/robots/editor/generated/robotsMetamodel.xml
	qrgui/elementEditor/shapeEdit/scene.cpp
	qrgui/elementEditor/shapeEdit/shapeEdit.cpp
	qrgui/elementEditor/shapeEdit/shapeEdit.h
	qrgui/elementEditor/shapeEdit/xmlLoader.cpp
	qrgui/elementEditor/shapeEdit/xmlLoader.h
	qrgui/mainwindow/mainWindow.cpp
	qrgui/mainwindow/mainWindow.h
	qrgui/pluginManager/editorManager.cpp
	qrgui/qrgui_ru.qm
	qrgui/umllib/element.h
	qrgui/umllib/nodeElement.cpp
	qrgui/umllib/nodeElement.h
	qrgui/umllib/sdfRenderer.cpp
	qrgui/umllib/sdfRenderer.h
	qrgui/view/editorViewScene.cpp
	qrgui/view/propertyEditorView.cpp
	qrutils/graphicsUtils/abstractItem.cpp
	qrutils/graphicsUtils/abstractScene.cpp
	qrutils/graphicsUtils/colorlisteditor.h
	qrutils/mathUtils/mathUtils.pri

+++ plugins/robots/robotsInterpreter/details/robotCommunication/robotCommunicationThreadBase.cpp
@@ -88,7 +88,8 @@ int RobotCommunicationThreadBase::i2cBytesReady(inputPort::InputPortEnum port)
 	QByteArray result;
 	send(command, lsGetStatusResponseSize, result);
 
-	if (result.isEmpty() || result[4] != errorCode::success) {
+	// static_cats<int> prevents a warning about operator != ambiguity
+	if (result.isEmpty() || static_cast<int>(result[4]) != errorCode::success) {
 		return 0;
 	} else {
 		return result[5];
